feat : PlayPopup 기능추가
1. WorkItem 을 통한 클릭 이벤트 정상 업그레이드 동작 확인
- 업그레이드 시 소지 금액 차감
- 업그레이드 시 레벨 상승
- 업그레이드 시 클릭 시 획득 금액 증가

2. LottoItem 기능 동작 확인
- Lotto 구매 시 소지 개수 증가 확인
- 소지한 Lotto가 있을 시 쿨타임 적용
- 쿨타임이 끝나면 Lotto 개봉
- 당첨 시 금액 확인

3. UI_Log 동작 확인
Lotto 개봉하여 당첨 된 경우 얼마만큼의 금액으로 당첨되었는지 로그 표시.

+++ I Will be a Rich/Assets/Scripts/Manager/GameManagerEx.cs
@@ -18,7 +18,7 @@ public class GameData
 	public int WorkAbility; // 화면 터치
 	public int[] LottoAbilities = new int[Define.MAX_LOTTO_COUNT];
 	public float[] LottoCoolTimes = new float[Define.MAX_LOTTO_COUNT];
-	public float[,] LottoProbability = new float[Define.MAX_LOTTO_COUNT, Define.LOTTO_PRIZE_COUNT];
+	public int[,] LottoProbability = new int[Define.MAX_LOTTO_COUNT, Define.LOTTO_PRIZE_COUNT];
     public int[] Upgrades = new int[Define.MAX_UPGRADE_COUNT];
 	public PlayerState Player;
 
@@ -63,7 +63,7 @@ public class GameManagerEx : MonoBehaviour
 		set { _gameData.LottoCoolTimes = value; }
 	}
 
-	public float[,] LottoProbability
+	public int[,] LottoProbability
     {
 		get { return _gameData.LottoProbability; }
         set { _gameData.LottoProbability = value; }
